requirements.txt
streamlit
google-genai
markdown
edge-tts